Modifying main to send unsigned short int

+++ main.c
@@ -3,7 +3,7 @@
 
 int main(void)
 {
-  int amounts[] = {1, 10, 39, 589, 2590, 31999};
+  unsigned short int amounts[] = {1, 10, 39, 589, 2590, 31999};
   for (int i = 0; i < 6; i++)
   {
     unsigned int notes = get_money(amounts[i]);
